Move related reader helpers to the tools subfolder
