+++ app/src/main/java/my/noveldokusha/ui/screens/reader/ReaderViewModel.kt
@@ -1,12 +1,9 @@
 package my.noveldokusha.ui.screens.reader
 
-import android.util.Log
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
 import androidx.lifecycle.*
-import com.google.mlkit.nl.translate.TranslateLanguage
-import com.google.mlkit.nl.translate.Translator
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.*
 import my.noveldokusha.AppPreferences
@@ -15,9 +12,10 @@ import my.noveldokusha.data.database.tables.Chapter
 import my.noveldokusha.tools.TranslationManager
 import my.noveldokusha.tools.TranslatorState
 import my.noveldokusha.ui.BaseViewModel
+import my.noveldokusha.ui.screens.reader.tools.ChaptersIsReadRoutine
+import my.noveldokusha.ui.screens.reader.tools.saveLastReadPositionState
 import my.noveldokusha.utils.StateExtra_String
 import java.io.File
-import java.util.Locale
 import java.util.concurrent.atomic.AtomicBoolean
 import javax.inject.Inject
 import kotlin.collections.ArrayList
@@ -40,15 +38,13 @@ class ReaderViewModel @Inject constructor(
     enum class ReaderState
     { IDLE, LOADING, INITIAL_LOAD }
 
+    data class ChapterState(val url: String, val position: Int, val offset: Int)
     data class ChapterStats(val itemCount: Int, val chapter: Chapter, val index: Int)
 
     override var bookUrl by StateExtra_String(state)
     override var chapterUrl by StateExtra_String(state)
 
-    var translator: TranslatorState? = translationManager.getTranslator(
-        source = TranslateLanguage.ENGLISH,
-        target = TranslateLanguage.CATALAN,
-    )
+    var translator: TranslatorState? = null
 
     val localBookBaseFolder = File(repository.settings.folderBooks, bookUrl.removePrefix("local://"))
 
@@ -65,7 +61,7 @@ class ReaderViewModel @Inject constructor(
     init
     {
         val chapter = viewModelScope.async(Dispatchers.IO) { repository.bookChapter.get(chapterUrl) }
-        val bookChapter = viewModelScope.async(Dispatchers.IO) { repository.bookChapter.chapters(bookUrl) }
+        val bookChapters = viewModelScope.async(Dispatchers.IO) { repository.bookChapter.chapters(bookUrl) }
 
         // Need to fix this somehow
         runBlocking {
@@ -74,7 +70,7 @@ class ReaderViewModel @Inject constructor(
                 position = chapter.await()?.lastReadPosition ?: 0,
                 offset = chapter.await()?.lastReadOffset ?: 0
             )
-            this@ReaderViewModel.orderedChapters = bookChapter.await()
+            this@ReaderViewModel.orderedChapters = bookChapters.await()
         }
     }
 
@@ -123,7 +119,7 @@ class ReaderViewModel @Inject constructor(
     suspend fun getChapterInitialPosition(): Pair<Int, Int>?
     {
         val stats = chaptersStats[currentChapter.url] ?: return null
-        return getChapterInitialPosition(
+        return my.noveldokusha.ui.screens.reader.tools.getChapterInitialPosition(
             repository = repository,
             bookUrl = bookUrl,
             chapter = stats.chapter,
@@ -131,55 +127,3 @@ class ReaderViewModel @Inject constructor(
         )
     }
 }
-
-data class ChapterState(val url: String, val position: Int, val offset: Int)
-
-private fun saveLastReadPositionState(
-    repository: Repository,
-    bookUrl: String,
-    chapter: ChapterState,
-    oldChapter: ChapterState? = null
-) = CoroutineScope(Dispatchers.IO).launch {
-    repository.withTransaction {
-        repository.bookLibrary.get(bookUrl)?.let {
-            repository.bookLibrary.update(it.copy(lastReadChapter = chapter.url))
-        }
-
-        if (oldChapter?.url != null) repository.bookChapter.get(oldChapter.url)?.let {
-            repository.bookChapter.update(it.copy(lastReadPosition = oldChapter.position, lastReadOffset = oldChapter.offset))
-        }
-
-        repository.bookChapter.get(chapter.url)?.let {
-            repository.bookChapter.update(it.copy(lastReadPosition = chapter.position, lastReadOffset = chapter.offset))
-        }
-    }
-}
-
-suspend fun getChapterInitialPosition(
-    repository: Repository,
-    bookUrl: String,
-    chapter: Chapter,
-    items: ArrayList<ReaderItem>
-): Pair<Int, Int> = coroutineScope {
-
-    val book = async(Dispatchers.IO) { repository.bookLibrary.get(bookUrl) }
-    val titlePos = async(Dispatchers.Default) {
-        items.indexOfFirst { it is ReaderItem.TITLE }
-    }
-    val position = async(Dispatchers.Default) {
-        items.indexOfFirst {
-            it is ReaderItem.Position && it.pos == chapter.lastReadPosition
-        }.let { index ->
-            if (index == -1) Pair(titlePos.await(), 0)
-            else Pair(index, chapter.lastReadOffset)
-        }
-    }
-
-    when
-    {
-        chapter.url == book.await()?.lastReadChapter -> position.await()
-        chapter.read -> Pair(titlePos.await(), 0)
-        else -> position.await()
-    }.let { Pair(it.first.coerceAtLeast(titlePos.await()), it.second) }
-}
-
